third_party/WebKit/LayoutTests/fast/canvas/gradient-addColorStop-with-invalid-offset-expected.txt

This test checks invalid offsets on gradients.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gradient.addColorStop(-1, '#000') threw exception IndexSizeError: Failed to execute 'addColorStop' on 'CanvasGradient': The provided value (-1) is outside the range (0.0, 1.0)..
PASS gradient.addColorStop(2, '#000') threw exception IndexSizeError: Failed to execute 'addColorStop' on 'CanvasGradient': The provided value (2) is outside the range (0.0, 1.0)..
PASS gradient.addColorStop(Infinity, '#000') threw exception TypeError: Failed to execute 'addColorStop' on 'CanvasGradient': The provided float value is non-finite..
PASS gradient.addColorStop(-Infinity, '#000') threw exception TypeError: Failed to execute 'addColorStop' on 'CanvasGradient': The provided float value is non-finite..
PASS gradient.addColorStop(NaN, '#000') threw exception TypeError: Failed to execute 'addColorStop' on 'CanvasGradient': The provided float value is non-finite..
PASS successfullyParsed is true

TEST COMPLETE
